# api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fixr Home Services API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Service {
  id: number;
  name: string;
  basePrice: number;
  description: string;
}

export type BookingUrgency =
  (typeof BookingUrgency)[keyof typeof BookingUrgency];

export const BookingUrgency = {
  normal: "normal",
  urgent: "urgent",
  emergency: "emergency",
} as const;

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Booking {
  id: number;
  service: string;
  description: string;
  address: string;
  city: string;
  urgency: BookingUrgency;
  status: BookingStatus;
  baseCost: number;
  urgencyFee: number;
  totalCost: number;
  workerId?: number | null;
  createdAt: string;
  completedAt?: string | null;
}

export type CreateBookingBodyUrgency =
  (typeof CreateBookingBodyUrgency)[keyof typeof CreateBookingBodyUrgency];

export const CreateBookingBodyUrgency = {
  normal: "normal",
  urgent: "urgent",
  emergency: "emergency",
} as const;

export interface CreateBookingBody {
  service: string;
  description: string;
  address: string;
  city: string;
  urgency: CreateBookingBodyUrgency;
}

export type UpdateBookingStatusBodyStatus =
  (typeof UpdateBookingStatusBodyStatus)[keyof typeof UpdateBookingStatusBodyStatus];

export const UpdateBookingStatusBodyStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateBookingStatusBody {
  status: UpdateBookingStatusBodyStatus;
  workerId?: number | null;
  finalCost?: number | null;
}

export type EstimateBookingCostBodyUrgency =
  (typeof EstimateBookingCostBodyUrgency)[keyof typeof EstimateBookingCostBodyUrgency];

export const EstimateBookingCostBodyUrgency = {
  normal: "normal",
  urgent: "urgent",
  emergency: "emergency",
} as const;

export interface EstimateBookingCostBody {
  service: string;
  urgency: EstimateBookingCostBodyUrgency;
  description?: string;
}

export interface CostEstimate {
  baseEstimate: number;
  urgencyFee: number;
  totalEstimate: number;
  urgencyApplied: boolean;
}

export interface Worker {
  id: number;
  name: string;
  initials: string;
  specialty: string;
  locationArea: string;
  locationCity: string;
  experienceYears: number;
  ratePerHour: number;
  skills: string[];
  rating: number;
  verified: boolean;
  avatarColor: string;
}

export type MaterialCategory =
  (typeof MaterialCategory)[keyof typeof MaterialCategory];

export const MaterialCategory = {
  plumbing: "plumbing",
  electrical: "electrical",
  hardware_tools: "hardware_tools",
  paint: "paint",
} as const;

export interface Material {
  id: number;
  name: string;
  brand: string;
  category: MaterialCategory;
  price: number;
  unit: string;
}

export interface AdminStats {
  totalRevenue: number;
  activeBookings: number;
  activeWorkers: number;
  pendingVerifications: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  booking: "booking",
  completion: "completion",
  worker: "worker",
} as const;

export interface ActivityItem {
  id: number;
  title: string;
  subtitle: string;
  type: ActivityItemType;
  createdAt: string;
}

export interface BookingsByService {
  service: string;
  count: number;
}

export type ListWorkersParams = {
  search?: string;
  specialty?: string;
};

export type ListMaterialsParams = {
  category?: string;
  search?: string;
};