# api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fixr Home Services API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: services
    description: Service categories
  - name: bookings
    description: Booking operations
  - name: workers
    description: Worker operations
  - name: materials
    description: Materials store
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all service categories
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"

  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingStatusBody"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  /bookings/estimate:
    post:
      operationId: estimateBookingCost
      tags: [bookings]
      summary: Get AI cost estimate for a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateBookingCostBody"
      responses:
        "200":
          description: Cost estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CostEstimate"

  /workers:
    get:
      operationId: listWorkers
      tags: [workers]
      summary: List/search workers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: specialty
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of workers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Worker"

  /workers/{id}:
    get:
      operationId: getWorker
      tags: [workers]
      summary: Get worker profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Worker profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"

  /materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List materials
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard stats
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

  /admin/activity:
    get:
      operationId: getRecentActivity
      tags: [admin]
      summary: Get recent platform activity
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  /admin/bookings-by-service:
    get:
      operationId: getBookingsByService
      tags: [admin]
      summary: Get booking counts grouped by service
      responses:
        "200":
          description: Bookings by service
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingsByService"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Service:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        basePrice:
          type: integer
        description:
          type: string
      required: [id, name, basePrice, description]

    Booking:
      type: object
      properties:
        id:
          type: integer
        service:
          type: string
        description:
          type: string
        address:
          type: string
        city:
          type: string
        urgency:
          type: string
          enum: [normal, urgent, emergency]
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        baseCost:
          type: integer
        urgencyFee:
          type: integer
        totalCost:
          type: integer
        workerId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, service, description, address, city, urgency, status, baseCost, urgencyFee, totalCost, createdAt]

    CreateBookingBody:
      type: object
      properties:
        service:
          type: string
        description:
          type: string
        address:
          type: string
        city:
          type: string
        urgency:
          type: string
          enum: [normal, urgent, emergency]
      required: [service, description, address, city, urgency]

    UpdateBookingStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        workerId:
          type: integer
          nullable: true
        finalCost:
          type: integer
          nullable: true
      required: [status]

    EstimateBookingCostBody:
      type: object
      properties:
        service:
          type: string
        urgency:
          type: string
          enum: [normal, urgent, emergency]
        description:
          type: string
      required: [service, urgency]

    CostEstimate:
      type: object
      properties:
        baseEstimate:
          type: integer
        urgencyFee:
          type: integer
        totalEstimate:
          type: integer
        urgencyApplied:
          type: boolean
      required: [baseEstimate, urgencyFee, totalEstimate, urgencyApplied]

    Worker:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        initials:
          type: string
        specialty:
          type: string
        locationArea:
          type: string
        locationCity:
          type: string
        experienceYears:
          type: integer
        ratePerHour:
          type: integer
        skills:
          type: array
          items:
            type: string
        rating:
          type: number
        verified:
          type: boolean
        avatarColor:
          type: string
      required: [id, name, initials, specialty, locationArea, locationCity, experienceYears, ratePerHour, skills, rating, verified, avatarColor]

    Material:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        brand:
          type: string
        category:
          type: string
          enum: [plumbing, electrical, hardware_tools, paint]
        price:
          type: integer
        unit:
          type: string
      required: [id, name, brand, category, price, unit]

    AdminStats:
      type: object
      properties:
        totalRevenue:
          type: integer
        activeBookings:
          type: integer
        activeWorkers:
          type: integer
        pendingVerifications:
          type: integer
      required: [totalRevenue, activeBookings, activeWorkers, pendingVerifications]

    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        subtitle:
          type: string
        type:
          type: string
          enum: [booking, completion, worker]
        createdAt:
          type: string
          format: date-time
      required: [id, title, subtitle, type, createdAt]

    BookingsByService:
      type: object
      properties:
        service:
          type: string
        count:
          type: integer
      required: [service, count]